Divide body content logic
