+++ website/frontend/src/components/pages/Explorer/EventsList/children/EventItem/children/Body/Body.module.scss
@@ -5,23 +5,10 @@
   padding-left: 30px;
 }
 
-.checkbox {
-  margin-bottom: 16px;
-}
-
 .pre {
   padding: 15px 25px;
   background-color: $mainBackground;
   border: 1px dashed $bkgButtonColor;
   border-radius: 2px;
   overflow-x: auto;
-}
-
-.line {
-  display: block;
-}
-
-.link {
-  text-decoration: underline;
-  color: inherit;
 }
