+++ website/frontend/src/components/pages/Explorer/EventsList/children/EventItem/children/Body/children/LogContent/LogContent.module.scss
@@ -1,0 +1,3 @@
+.checkbox {
+  margin-bottom: 16px;
+}
